// web/src/api/generated/model/confirmIdIn.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface ConfirmIdIn {
  /** @maxLength 32 */
  id_number: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  surname: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  first_names: string;
  date_of_birth: string;
}